requirements.txt
fastapi>=0.121.2
uvicorn>=0.38.0
openai>=1.0.0
pydantic-settings>=2.2.1
python-dotenv>=1.0.0
